scripts/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
requests==2.32.3
python-multipart==0.0.20
Pillow==11.0.0
python-dotenv==1.0.1
